Reapply "fix: rollback to use mock orders"
This reverts commit 72a8d354951299f4c5e6352409b765a20683770d.
