+++ packages/marketplace/ui/src/lib/hooks/use-order.tsx
@@ -1,7 +1,23 @@
-import { useOrders } from "./use-orders";
+export const useOrder = ({ id }: { id: bigint }) => {
+  // TODO: unmock
 
-export function useOrder(orderId: bigint) {
-  const { sequencerRollups, replicaRollups, isLoading, error, refetch } = useOrders();
-  const order = [...sequencerRollups, ...replicaRollups].find(o => o.id === orderId);
-  return { order, isLoading, error, refetch };
-}
+  return {
+    name: "mock rollup",
+    id: 1n,
+    fulfilledAt: 0n,
+    terminatedAt: 0n,
+    provider: {
+      sequencer: "0x123",
+      batcher: "0x123",
+      proposer: "0x123",
+      challenger: "0x123",
+    },
+    network: { l1ChainId: 1n },
+    addresses: {
+      systemConfigProxy: "0x123",
+      l2OutputOracleProxy: "0x123",
+      systemOwnerSafe: "0x123",
+      proxyAdmin: "0x123",
+    },
+  };
+};
