+++ packages/marketplace/ui/src/routes/rollups/$id.lazy.tsx
@@ -8,6 +8,7 @@ import { Button } from "@/components/ui/button";
 import { Clock, Download, OctagonAlert } from "lucide-react";
 import { cn } from "@/lib/utils";
 import { useOrder } from "@/lib/hooks/use-order";
+import { Order } from "@/types";
 
 export const Route = createLazyFileRoute("/rollups/$id")({
   component: RollupDashboard,
@@ -16,11 +17,9 @@ export const Route = createLazyFileRoute("/rollups/$id")({
 export default function RollupDashboard() {
   const { id } = Route.useParams();
   const [statusColor, setStatusColor] = useState("gray-200");
-  const { order, isLoading } = useOrder(BigInt(id))
-
-  if (isLoading) return <></>;
-  if (!order) return <p>Order not found</p>;
-  const { terminatedAt, fulfilledAt} = order
+  const { terminatedAt, fulfilledAt, name } = useOrder({id: BigInt(id)})
+  const order = { terminatedAt, fulfilledAt, name, id }
+  if (!terminatedAt) return <p>Order not found</p>;
   return (
     <div className="md:p-6 space-y-6">
       <div
@@ -34,10 +33,10 @@ export default function RollupDashboard() {
             terminatedAt > 0n ? "border border-alert-border" : "border-gray-200"
           )}
         >
-          <RollupHeader order={order} />
+          <RollupHeader order={order as unknown as Order} />
             {fulfilledAt > 0n && (
               <RollupActions
-                order={order}
+                order={order as unknown as Order}
                 setStatusColor={setStatusColor}
                 statusColor={statusColor}
               />
